Create a directory for KONAMI and moves its roms to it
This is a first step to create a common directory structure
for the games. Instead of having two different directories
(one for ROMS and another for DSK), is better to have only
one directory for each company. The base directory is hashed
with the first letter of the company because it helps because
the number of companies will be large.