Add BANK BUSTER
This game is not working due to the fact it needs too much RAM, so it is
not possible to execute it from the hard disk, but it doesn't work either
from the floppy emulation. I can try to see later why the disk emulation
is not giving enough memory.