Move Dungeon Striker to dsk file
This game can be run in a common MSX-DOS filesystem, but it uses names
that are non valid in a MS-DOS filesystem, so it creates problems with
the repository. This is a game that people is not going to play a lot,
so put it in dsk format and forgive the problem.